refactor(Events): ♻️ Move one folder up
